Fix & update ssh vm setup & instructions
Note that the regeneration of the SSH host keys was failing because the
dpkg-reconfigure command showed an interactive prompt to ask which
version of the configuration to use (the one that came with the package
or the modified local version). The addition of the DEBIAN_FRONTEND
environment variable solved the issue.